add dynamic generator test

+++ mybatis-r2dbc/src/test/java/pro/chenggang/project/reactive/mybatis/support/r2dbc/application/mapper/DeptMapper.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @Mapper
 public interface DeptMapper {
 
-    Mono<Integer> count();
+    Mono<Long> count();
 
     Flux<Dept> selectAll();
